Added --add-repo. Added --rootdir. Recognizing (srpm-)buildroot profiles.
To enable modular dependencies we need to add the --add-repo argument so
we can add other repositories to the buildroot which contain modular
dependencies.
We needed to add the --rootdir argument, so we can specify the chroot
dir for the mock tool. Without this we could ran out of disk space for
some builds.
(srpm-)buildroot profiles are now recognized in modular dependencies.
They are enabled and installed in the (srpm-)buildroot before the build starts.

Signed-off-by: Martin Curlej <mcurlej@redhat.com>